+++ src/main/java/seedu/address/storage/XmlSerializableAddressBook.java
@@ -11,6 +11,7 @@ import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.person.Person;
+import seedu.address.storage.adapter.XmlAdaptedPerson;
 
 /**
  * An Immutable AddressBook that is serializable to XML format
