+++ src/test/java/seedu/address/storage/XmlAdaptedPersonTest.java
@@ -1,7 +1,7 @@
 package seedu.address.storage;
 
 import static org.junit.Assert.assertEquals;
-import static seedu.address.storage.XmlAdaptedPerson.MISSING_FIELD_MESSAGE_FORMAT;
+import static seedu.address.storage.adapter.XmlAdaptedPerson.MISSING_FIELD_MESSAGE_FORMAT;
 import static seedu.address.testutil.TypicalPersons.BENSON;
 
 import java.util.ArrayList;
@@ -15,6 +15,8 @@ import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
+import seedu.address.storage.adapter.XmlAdaptedPerson;
+import seedu.address.storage.adapter.XmlAdaptedTag;
 import seedu.address.testutil.Assert;
 
 public class XmlAdaptedPersonTest {
